% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGenes.r
\name{annotateGenes}
\alias{annotateGenes}
\title{Annotate genomic position using bumphunter's matchGenes}
\usage{
annotateGenes(DMS)
}
\arguments{
\item{DMS}{obj w/ chr, start, end positions \code{DMS}}
}
\value{
annotated data frame; see bumphunter documentation for further 
explanation of output
}
\description{
Annotate genomic position using bumphunter's matchGenes
}
\examples{
DMS_annotated=annotateGenes(DMS)
}
\keyword{annotate,}
\keyword{bumphunter,}
\keyword{matchGenes}

