% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Boxplot of single sites from output from DMS Analysis}
\usage{
bootstrap(methylKit.obj, covariates, variance, JOBID)
}
\arguments{
\item{methylKit.obj}{output methylkit object \code{methylKit.obj}}

\item{covariates}{covariate matrix \code{covariates}}

\item{variance}{what proportion of top variant sites to include \code{}}
}
\value{
Rdata object for bootstrap
}
\description{
Boxplot of single sites from output from DMS Analysis
}
\examples{
bootstrap(meth_normalized_N63_count100_20L, covariates=covariates_N63, variance=0.75,JOBID=2)
}
\keyword{DMS}
\keyword{boostrap,}
\keyword{limma,}

