% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_array.R
\name{bootstrap_array}
\alias{bootstrap_array}
\title{Boxplot of single sites from output from DMS Analysis}
\usage{
bootstrap_array(CpG, covariates, variance, JOBID)
}
\arguments{
\item{CpG}{DNAm data from array \code{CpG}}

\item{covariates}{covariate matrix \code{covariates}}

\item{variance}{what proportion of top variant sites to include \code{}}
}
\value{
Rdata object for bootstrap
}
\description{
Boxplot of single sites from output from DMS Analysis
}
\examples{
bootstrap_array(CpG, covariates=covariates, variance=0.75,JOBID=2)
}
\keyword{DMS}
\keyword{boostrap,}
\keyword{limma,}

