% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_noout.R
\name{bootstrap_noout}
\alias{bootstrap_noout}
\title{Boxplot of single sites from output from DMS Analysis}
\usage{
bootstrap_noout(methylKit.obj, covariates)
}
\arguments{
\item{methylKit.obj}{output methylkit object \code{methylKit.obj}}

\item{covariates}{covariate matrix \code{covariates}}
}
\value{
bootstrapped pvalues in current R environment
}
\description{
Boxplot of single sites from output from DMS Analysis
}
\examples{
bootstrap(meth_normalized_N63_count100_20L, covariates=covariates_N63)
}
\keyword{DMS}
\keyword{boostrap,}
\keyword{limma,}

