% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_DMR.R
\name{boxplot_DMR}
\alias{boxplot_DMR}
\title{Boxplot of regions from output from DMR Analysis}
\usage{
boxplot_DMR(DMR, methyl.mat, treatment, filename, n = 15)
}
\arguments{
\item{methyl.mat}{matrix with methylation values \code{methyl.mat}}

\item{treatment}{vector with case control status \code{treatment}}

\item{n}{number of sites you'd like to output \code{n}}

\item{annotated}{DMR table output from runbumphunter  \code{DMR}}
}
\value{
plot single site analysis output
}
\description{
Boxplot of regions from output from DMR Analysis
}
\examples{
boxplot_DMR(DMR,methyl.mat,treatment,filename,n=15)
}
\keyword{DMR}
\keyword{boxplot,}
\keyword{bumphunter,}

