% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_DMS.R
\name{boxplot_DMS}
\alias{boxplot_DMS}
\title{Boxplot of single sites from output from DMS Analysis}
\usage{
boxplot_DMS(DMS, methyl.mat, treatment, filename, n = 15)
}
\arguments{
\item{DMS}{output dataframe produced by runlimma \code{DMS}}

\item{methyl.mat}{matrix with methylation values \code{methyl.mat}}

\item{treatment}{vector with case control status \code{treatment}}

\item{n}{number of sites you'd like to output \code{n}}
}
\value{
plot single site analysis output
}
\description{
Boxplot of single sites from output from DMS Analysis
}
\examples{
boxplot_DMS(DMS,methyl.mat,treatment,filename,n=15)
}
\keyword{boxplot,}
\keyword{limma,}

