% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_VMR.R
\name{boxplot_VMR}
\alias{boxplot_VMR}
\title{Boxplot of regions from output from VMR Analysis}
\usage{
boxplot_VMR(VMR, methyl.mat, treatment, filename, n = 15)
}
\arguments{
\item{VMR}{table output from VMR_Finder (Andrew Jaffe code modified)  \code{VMR}}

\item{methyl.mat}{matrix with methylation values \code{methyl.mat}}

\item{treatment}{vector with case control status \code{treatment}}

\item{n}{number of sites you'd like to output \code{n}}
}
\value{
plot single site analysis output
}
\description{
Boxplot of regions from output from VMR Analysis
}
\examples{
boxplot_VMR(VMR,methyl.mat,treatment,filename,n=15)
}
\keyword{VMR}
\keyword{boxplot,}
\keyword{bumphunter,}

