% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circos.R
\name{Circos}
\alias{Circos}
\title{QQ plot of all sites from DMS Analysis}
\usage{
Circos(DMS, methdiff = 5, pval = 0.01, hyper.col = "springgreen3",
  hypo.col = "mediumvioletred")
}
\arguments{
\item{DMS}{output dataframe produced by runlimma \code{DMS}}

\item{methdiff}{difference in methylation \code{methdiff}}

\item{pval}{pvalue cutoff \code{pval}}
}
\value{
QQplot single site analysis output
}
\description{
QQ plot of all sites from DMS Analysis
}
\examples{
Circos(DMS,methdiff=10,pval=0.05,
hyper.col="springgreen3",hypo.col="mediumvioletred")
}
\keyword{QQplot}
\keyword{limma,}

