% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieGraph.R
\name{CpGPie}
\alias{CpGPie}
\title{Pie Chart: CGI Annotation}
\usage{
CpGPie(methylKit.obj, cpg.obj)
}
\arguments{
\item{methylKit.obj}{obj with chr,start,end positions \\code{methylKit.obj}}

\item{cpg.obj}{SimpleGenomicRangesList w/ CGI info \code{cpg.obj}}
}
\value{
Pie Chart for CGI annotation
}
\description{
Pie Chart: CGI Annotation
}
\examples{
CpGPie(methylKit.obj,cpg.obj)
}
\keyword{PieChart,CGI,}
\keyword{annotation}
\keyword{methylBase,}
\keyword{methylKit,}

