% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePie.R, R/PieGraph.R
\name{GenePie}
\alias{GenePie}
\title{Pie Chart: GeneAnnotation}
\usage{
GenePie(methylKit.obj, gene.obj)

GenePie(methylKit.obj, gene.obj)
}
\arguments{
\item{methylKit.obj}{methylBase object \code{methylKit.obj}}

\item{gene.obj}{GRaangesList w/ promoter,exon,intron,and TSS \code{gene.obj}}

\item{methylKit.obj}{obj w/ chr, start, end positions \code{methylKit.obj}}

\item{gene.obj}{GRaangesList w/ promoter,exon,intron,and TSS \code{gene.obj}}
}
\value{
Pie Chart for genic annotation

Pie Chart for genic annotation
}
\description{
Pie Chart: GeneAnnotation

Pie Chart: GeneAnnotation
}
\examples{
pie<-GenePie(methylBase,gene.obj)
plot(pie)
GenePie(methylKit.obj,gene.obj)
}
\keyword{PieChart,}
\keyword{annotation}
\keyword{methylBase,}
\keyword{methylKit,}

