% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runlimma.R
\name{hyper}
\alias{hyper}
\title{Get Hypermethylated Sites}
\usage{
hyper(DMS, methdiff = 0, pval = NULL)
}
\arguments{
\item{DMS}{output from DMS Analysis \code{DMS}}

\item{methdiff}{difference in methylation \code{methdiff}}

\item{pval}{pvalue cutoff \code{pval}}
}
\value{
data frame with hypermethylated sites
}
\description{
Get Hypermethylated Sites
}
\examples{
DMS_hyper <- hyper(DMS,methdiff=0,pval=1)
}
\keyword{hypermethylation}
\keyword{methylBase,hyper,}

