% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runlimma.R
\name{hypo}
\alias{hypo}
\title{Get Hypomethylated Sites}
\usage{
hypo(DMS, methdiff = 0, pval = NULL)
}
\arguments{
\item{DMS}{output from DMS Analysis \code{DMS}}

\item{methdiff}{difference in methylation \code{methdiff}}

\item{pval}{pvalue cutoff \code{pval}}
}
\value{
data frame with hypomethylated sites
}
\description{
Get Hypomethylated Sites
}
\examples{
 DMS_hypo <- hypo(DMS,methdiff=0,pval=1)
}
\keyword{hypomethylation}
\keyword{methylBase,hypo,}

