% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Karyogram.R
\name{karyo}
\alias{karyo}
\title{Karyogram:}
\usage{
karyo(methylKit.obj, cols)
}
\arguments{
\item{methylKit.obj}{object with chr,start,end positions \code{methylKit.obj}}

\item{cols}{color vector \code{cols}}
}
\value{
Karyogram with CpGs plotted across the genome (chr1-22)
}
\description{
Karyogram:
}
\examples{
karyogram<-karyo(methylBase,cols="limegreen")
plot(karyogram)
}
\keyword{Karyogram,}
\keyword{karyo}
\keyword{methylBase,}
\keyword{methylKit,}

