% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan plot of all sites from DMS Analysis}
\usage{
manhattan(DMS, filename, sig = NULL)
}
\arguments{
\item{DMS}{output dataframe produced by runlimma \code{DMS}}

\item{filename}{filenames to attach \code{filename}}

\item{sig}{level of genome-wide significance \code{sig}}
}
\value{
plot single site analysis output
}
\description{
Manhattan plot of all sites from DMS Analysis
}
\examples{
manhattan(DMS,"test")
}
\keyword{limma,}
\keyword{manhattan,}

