% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{Null \code{plotlist}}

\item{file}{FILE \code{file}}

\item{cols}{Number of columns \code{cols}}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
allow ggplot to multipanel plot
}
\description{
Multiplot function
}
\examples{
multiplot(p1,p2,p3, cols=3)
}
\keyword{DMR}
\keyword{multiplot,}

