% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQplot.R
\name{obs}
\alias{obs}
\title{QQ plot of all sites from DMS Analysis broken down by Mean Methylation}
\usage{
obs(pval)
}
\arguments{
\item{DMS}{output dataframe produced by runlimma \code{DMS}}

\item{filename}{filenames to attach \code{filename}}
}
\value{
QQplot single site analysis output
}
\description{
QQ plot of all sites from DMS Analysis broken down by Mean Methylation
}
\examples{
QQplot(DMS,"test")
}
\keyword{QQplot}
\keyword{limma,}

