% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runbumphunter.R
\name{runbumphunter}
\alias{runbumphunter}
\title{Run bumphunter to obtain DMRs (parameters stolen straight from bumphunter)}
\usage{
runbumphunter(methyl.mat, treatment, covariates, pickCutoff = TRUE,
  pickCutoffQ = 0.975, maxGap = 300, smooth = FALSE, B = 250,
  nullMethod = "bootstrap", n.core = 20)
}
\arguments{
\item{methyl.mat}{matrix of samples by methylation \code{DMS}}

\item{covariates}{matrix (samples in rows, covar in column) \code{covariates}}

\item{pickCutoff}{Should bumphunter attempt to pick a cutoff using the 
permutation distribution?}

\item{pickCutoffQ}{The quantile used for picking the cutoff using the 
permutation distribution.}

\item{maxGap}{If cluster is not provided this maximum location gap will 
be used to define cluster via the clusterMaker function.}

\item{smooth}{A logical value. If TRUE the estimated profile will be 
smoothed with the smoother defined by smoothFunction}

\item{B}{An integer denoting the number of resamples to use when computing 
null distributions. This defaults to 0. If permutations is supplied that 
defines the number of permutations/bootstraps and B is ignored. \code{B}}

\item{nullMethod}{Method used to generate null candidate regions, must be 
one of ‘bootstrap’ or ‘permutation’ (defaults to ‘permutation’). However, 
if covariates in addition to the outcome of interest are included in the 
design matrix (ncol(design)>2), the ‘permutation’ approach is not recommended.}

\item{n.core}{number of cores to run in paralell}
}
\value{
annotated data frame; see bumphunter documentation for further 
explanation of output
}
\description{
Run bumphunter to obtain DMRs (parameters stolen straight from bumphunter)
}
\examples{
DMS_annotated=annotateGenes(DMS)
}
\keyword{DMRs}
\keyword{bumphunter,}

