% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDVS.R
\name{runDVS}
\alias{runDVS}
\title{DVS Analysis}
\usage{
runDVS(methylKit.obj, covariates)
}
\arguments{
\item{methylKit.obj}{methylBase object \code{methylKit.obj}}

\item{covariates}{covariate data frame \code{covariates}}

\item{variance}{cutoff for sites to use based on variance \code{variance}

return single site DVS analysis (using limma's lmFit) 
adjusts for covariates
and weights methylation values based on log10(coverage)}
}
\description{
DVS Analysis
}
\examples{
DVS<- runDVS(methylkit.obj,covariates=covariates_N63,variance=0.75)
}
\keyword{DVS,}
\keyword{limma,}
\keyword{methylBase,}
\keyword{methylKit,}
\keyword{variance}

