% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runlimma.R
\name{runlimma}
\alias{runlimma}
\title{DMS Analysis}
\usage{
runlimma(methylKit.obj, covariates, variance)
}
\arguments{
\item{methylKit.obj}{methylBase object \code{methylKit.obj}}

\item{covariates}{covariate data frame \code{covariates}}

\item{variance}{cutoff for sites to use based on variance \code{variance}}
}
\value{
single site analysis (using limma's lmFit) -- adjusts for covariates
and weights methylation values based on log10(coverage)
}
\description{
DMS Analysis
}
\examples{
DMS<- runlimma(methylkit.obj,covariates,variance=0.75)
}
\keyword{DMS,}
\keyword{limma,}
\keyword{methylBase,}
\keyword{methylKit,}
\keyword{variance}

