% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitesperchr.R
\name{sitesperchr}
\alias{sitesperchr}
\title{Plot number of sites by chromosome}
\usage{
sitesperchr(DMS, pval)
}
\arguments{
\item{DMS}{output dataframe produced by runlimma \code{DMS}}

\item{pval}{filenames to attach \code{pval}}
}
\value{
QQplot single site analysis output
}
\description{
Plot number of sites by chromosome
}
\examples{
sitesperchr(DMS,pval=pvalue)
}
\keyword{QQplot}
\keyword{limma,}

