% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{Volcano plot of all sites from DMS Analysis}
\usage{
volcano(DMS, filename)
}
\arguments{
\item{DMS}{output dataframe produced by runlimma \code{DMS}}
}
\value{
volcano plot for single site analysis output
}
\description{
Volcano plot of all sites from DMS Analysis
}
\examples{
volcano(DMS,"test")
}
\keyword{limma,}
\keyword{volcano}

